int32 fix

+++ HW2-3/big_integer.cpp
@@ -4,11 +4,11 @@
 #include <cstring>
 #include <stdexcept>
 
-unsigned int ones = UINT32_MAX;
-unsigned long long base = 1ll + ones;
+uint32_t ones = UINT32_MAX;
+uint64_t base = 1ll + ones;
 int weight = 32;
 
-//unsigned long long base = static_cast<unsigned long long>(1000);
+//uint64_t base = static_cast<uint64_t>(1000);
 
 big_integer::big_integer()
 {
@@ -22,14 +22,14 @@ big_integer::big_integer(big_integer const& other)
     digits = other.digits;
 }
 
-big_integer::big_integer(long long a) {
+big_integer::big_integer(int64_t a) {
     sign = (a >= 0 ? 1 : -1);
-    auto num = static_cast<unsigned long long>(a * sign);
+    auto num = static_cast<uint64_t>(a * sign);
     while (num >= base) {
-        digits.push_back(static_cast<unsigned int>(num % base));
+        digits.push_back(static_cast<uint32_t>(num % base));
         num /= base;
     }
-    digits.push_back(static_cast<unsigned int>(num));
+    digits.push_back(static_cast<uint32_t>(num));
 }
 
 big_integer::big_integer(std::string const& str)
@@ -42,7 +42,7 @@ big_integer::big_integer(std::string const& str)
                 sign = -1;
         }
         else {
-            *this += mul_short_abs(pow, static_cast<unsigned int>(str[i] - '0'));
+            *this += mul_short_abs(pow, static_cast<uint32_t>(str[i] - '0'));
             pow = mul_short_abs(pow, 10);
         }
     }
@@ -147,14 +147,14 @@ big_integer& big_integer::operator<<=(int rhs)
 {
     int offset = rhs / weight;
     rhs %= weight;
-    this->digits.resize(static_cast<unsigned int>(this->digits.size() + offset + 1));
+    this->digits.resize(static_cast<uint32_t>(this->digits.size() + offset + 1));
     for (size_t i = this->digits.size() - offset - 2; i < this->digits.size(); i--) {
-        unsigned long long shift = 1ull * (*this)[i] << rhs;
+        uint64_t shift = 1ull * (*this)[i] << rhs;
         if (shift >= base) {
-            (*this)[i + offset + 1] += static_cast<unsigned int>(shift >> weight);
+            (*this)[i + offset + 1] += static_cast<uint32_t>(shift >> weight);
             shift &= ones;
         }
-        (*this)[i + offset] = static_cast<unsigned int>(shift);
+        (*this)[i + offset] = static_cast<uint32_t>(shift);
     }
     for (int i = 0; i < offset; i++) {
         (*this)[i] = 0;
@@ -168,8 +168,8 @@ big_integer& big_integer::operator>>=(int rhs)
     int offset = (rhs + weight - 1) / weight;
     rhs %= weight;
     (*this)[0] >>= rhs;
-    unsigned int one = 1;
-    unsigned int rhs_ones = 0;
+    uint32_t one = 1;
+    uint32_t rhs_ones = 0;
     for (int i = 0; i < rhs; i++) {
         rhs_ones += one;
         one <<= 1;
@@ -179,7 +179,7 @@ big_integer& big_integer::operator>>=(int rhs)
     }
 
     for (int i = offset; i < (int)this->digits.size(); i++) {
-        unsigned int left = (*this)[i] & rhs_ones;
+        uint32_t left = (*this)[i] & rhs_ones;
         (*this)[i] >>= rhs;
         (*this)[i - offset] += left << (32 - rhs);
     }
@@ -301,7 +301,7 @@ bool operator<(big_integer const& a, big_integer const& b)
 {
     if (a.sign != b.sign) return a.sign < b.sign;
     if (a.digits.size() != b.digits.size()) {
-        return a.sign * ((int)a.digits.size() - (int)b.digits.size()) < 0;
+        return a.sign * (a.digits.size() - b.digits.size()) < 0;
     }
 
     for (size_t i = a.digits.size() - 1; i < a.digits.size(); i--) {
@@ -331,10 +331,10 @@ std::string to_string(big_integer const& a)
     std::string res;
     if (a == 0) res = "0";
     big_integer tmp = a;
-    int dec_base = 100000000;
+    int32_t dec_base = 100000000;
     int zeroes = 8;
     do {
-        int mod = big_integer::cast_to_uint(tmp % dec_base);
+        int32_t mod = big_integer::cast_to_uint(tmp % dec_base);
         if (tmp.digits.size() > 1) {
             for (int i = 0; i < zeroes; i++) {
                 res += ('0' + mod % 10);
@@ -360,14 +360,14 @@ std::ostream& operator<<(std::ostream& s, big_integer const& a)
     return s << to_string(a);
 }
 
-unsigned int &big_integer::operator[](int const& index) {
-    return reinterpret_cast<unsigned int &>(this->digits[index]);
+uint32_t &big_integer::operator[](int const& index) {
+    return reinterpret_cast<uint32_t &>(this->digits[index]);
 }
-const unsigned int big_integer::operator[](int const& index) const {
+const uint32_t big_integer::operator[](int const& index) const {
     return this->digits[index];
 }
 
-unsigned int big_integer::cast_to_uint(big_integer const &a) {
+uint32_t big_integer::cast_to_uint(big_integer const &a) {
     return a[0];
 }
 
@@ -386,7 +386,7 @@ big_integer big_integer::add_abs(big_integer const &a, big_integer const &b) {
     loc_b.digits.resize(len);
 
     for (size_t i = 0; i < len; i++) {
-        unsigned long long tmp_sum = 0ull + loc_a[i] + loc_b[i] + carry;
+        uint64_t tmp_sum = 0ull + loc_a[i] + loc_b[i] + carry;
         if (tmp_sum >= base) {
             carry = 1;
             tmp_sum &= ones;
@@ -394,7 +394,7 @@ big_integer big_integer::add_abs(big_integer const &a, big_integer const &b) {
         else {
             carry = 0;
         }
-        sum[i] = static_cast<unsigned int>(tmp_sum);
+        sum[i] = static_cast<uint32_t>(tmp_sum);
     }
     if (carry != 0) sum.digits.push_back(carry);
     return sum;
@@ -414,10 +414,10 @@ big_integer big_integer::sub_abs(big_integer const &a, big_integer const &b) {
 
     for (size_t i = 0; i < loc_a.digits.size(); i++) {
 
-        unsigned int tmp_sum;
+        uint32_t tmp_sum;
         if (loc_a[i] < loc_b[i]) {
             loc_a[i + 1]--;
-            tmp_sum = static_cast<unsigned int>(base + loc_a[i] - loc_b[i]);
+            tmp_sum = static_cast<uint32_t>(base + loc_a[i] - loc_b[i]);
         }
         else {
             tmp_sum = loc_a[i] - loc_b[i];
@@ -434,11 +434,11 @@ big_integer big_integer::mul_abs(big_integer const &a, big_integer const &b) {
     ans.digits.resize(a.digits.size() + b.digits.size() + 1);
 
     for (size_t i = 0; i < a.digits.size(); i++) {
-        unsigned int carry = 0;
+        uint32_t carry = 0;
         for (size_t j = 0; j < b.digits.size(); j++) {
-            unsigned long long tmp_mul = 1ull * a[i] * b[j] + carry + ans[i + j];
-            carry = static_cast<unsigned int>(tmp_mul >> weight);
-            ans[i + j] = static_cast<unsigned int>(tmp_mul & ones);
+            uint64_t tmp_mul = 1ull * a[i] * b[j] + carry + ans[i + j];
+            carry = static_cast<uint32_t>(tmp_mul >> weight);
+            ans[i + j] = static_cast<uint32_t>(tmp_mul & ones);
         }
         ans[i + b.digits.size()] += carry;
     }
@@ -446,21 +446,21 @@ big_integer big_integer::mul_abs(big_integer const &a, big_integer const &b) {
     return ans;
 }
 
-big_integer big_integer::mul_short_abs(big_integer const &a, unsigned int b) {
+big_integer big_integer::mul_short_abs(big_integer const &a, uint32_t b) {
     big_integer ans;
-    unsigned int carry = 0;
+    uint32_t carry = 0;
     ans.digits.resize(a.digits.size() + 1);
     for (size_t i = 0; i < a.digits.size(); i++) {
-        unsigned long long tmp_mul = 1ull * a[i] * b + carry;
-        carry = static_cast<unsigned int>(tmp_mul >> weight);
-        ans[i] = static_cast<unsigned int>(tmp_mul & ones);
+        uint64_t tmp_mul = 1ull * a[i] * b + carry;
+        carry = static_cast<uint32_t>(tmp_mul >> weight);
+        ans[i] = static_cast<uint32_t>(tmp_mul & ones);
     }
     ans[a.digits.size()] = carry;
     shorten(ans);
     return ans;
 }
 
-bool big_integer::cmp_prefix(big_integer const &r, big_integer const &d, unsigned int k, unsigned int m)
+bool big_integer::cmp_prefix(big_integer const &r, big_integer const &d, uint32_t k, uint32_t m)
 {
     size_t i = m, j = 0;
     while (i != j) {
@@ -474,7 +474,7 @@ std::pair<big_integer, big_integer> big_integer::div_abs(big_integer const& a, b
     big_integer loc_a = a, loc_b = b;
     big_integer r, q;
     size_t n = loc_a.digits.size(), m = loc_b.digits.size();
-    unsigned int f = static_cast<unsigned int>(base / (loc_b[m - 1] + 1));
+    uint32_t f = static_cast<uint32_t>(base / (loc_b[m - 1] + 1));
 
     r = (f == 0) ? (loc_a << weight) : (mul_short_abs(loc_a, f));
     big_integer const d = (f == 0) ? (loc_b << weight) : (mul_short_abs(loc_b, f));
@@ -483,15 +483,15 @@ std::pair<big_integer, big_integer> big_integer::div_abs(big_integer const& a, b
     r.digits.push_back(0);
     big_integer dq;
     for (size_t k = n - m; k < n - m + 1; k--) {
-        unsigned int qt;
+        uint32_t qt;
         size_t km = k + m;
         if (r == 0) {
             qt = 0;
         }
         else {
-            unsigned long long r1 = ((1ull * r[km]) << weight) + r[km - 1];
-            unsigned long long d1 = d[m - 1];
-            qt = std::min(static_cast<unsigned int>(r1 / d1), ones);
+            uint64_t r1 = ((1ull * r[km]) << weight) + r[km - 1];
+            uint64_t d1 = d[m - 1];
+            qt = std::min(static_cast<uint32_t>(r1 / d1), ones);
         }
         if (qt == 0)
             continue;
@@ -511,17 +511,17 @@ std::pair<big_integer, big_integer> big_integer::div_abs(big_integer const& a, b
     return {q, r};
 }
 
-std::pair<big_integer, unsigned int> big_integer::div_short_abs(big_integer const& a, unsigned int b) {
+std::pair<big_integer, uint32_t> big_integer::div_short_abs(big_integer const& a, uint32_t b) {
     big_integer div;
     div.digits.resize(a.digits.size());
-    unsigned long long mod = 0;
+    uint64_t mod = 0;
     for (size_t i = a.digits.size() - 1; i < a.digits.size(); i--) {
-        unsigned long long cur = a[i] + (mod << weight);
-        div[i] = static_cast<unsigned int>(cur / b);
+        uint64_t cur = a[i] + (mod << weight);
+        div[i] = static_cast<uint32_t>(cur / b);
         mod = cur % b;
     }
     shorten(div);
-    return {div, static_cast<unsigned int>(mod)};
+    return {div, static_cast<uint32_t>(mod)};
 }
 
 big_integer big_integer::bitwise(big_integer a, big_integer b, char op) {
